restructure api module into subpackage
